requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
